requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zstandard==0.23.0
pytest  # For running tests
